THEME: add upload to theme for homepage hero image
Added a field to the cfbtv theme settings so you can eaisly select
the image you want on the homepage.  This turned out to be a little
tricky because Drupal doesn't automatically save the image file
perminently if it is not used in a node (this if for the theme
remember).  The additional function "_permanently_save_image()"
in theme_settings.php covers this.