MODULES: allows me to auto "whiten" page banners
the imagecache_actions module adds a whole bunch more options for
creating drupal "image styles", including a color overlay.  Using
this, I can automatically stylize any image that is chosen for the
page banner.

this module does require imagemagick to be installed on the host
machine.  the imagemagick module helps drupal utilize imagemagick
but you must install the actual tool yourself - Drupal can't do that
for you. read the drupal project page carefully if there is trouble
with this module.