FILES: unignoring files directories for site imgs
In order to make it easier to replicate the site on dev/stage/prod
with all the admin-defined images, I've started tracking the files
directories that store those images.

I'm still ignoring most of the other user & module content stored
in /sites/default/files. (Per normal drupal convention, that content
should not be vital to the site).

Down the road it might be a better idea to relocate the directories
for node & theme images to some other location, but I'm not sure
where I want to put it yet.